pyproject.toml
[project]
name = "scraping_coto"
version = "0.0.1"
authors = [{ name = "josenilton1878@gmail.com" }]
requires-python = ">= 3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "bs4>=0.0.2",
    "databricks-bundles==0.276.0",
    "databricks-connect>=16.1.6",
    "delta-spark>=3.3.2",
    "dotenv>=0.9.9",
    "ipykernel>=6.30.1",
    "logfire[pydantic,requests,system-metrics]>=4.3.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=0.30.1",
    "nest-asyncio>=1.6.0",
    "playwright>=1.56.0",
    "polars>=1.32.3",
    "pydantic>=2.11.7",
    "python-dotenv[cli]>=1.1.1",
    "pytz>=2025.2",
    "random-user-agent>=1.0.1",
    "requests>=2.32.4",
    "taskipy>=1.14.1",
    "unidecode>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["uv_build>=0.8.11"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pandas>=2.3.1",
    "pytest>=8.4.1",
]
docs = [
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-literate-nav>=0.6.2",
]

[tool.uv.sources]
scraping_coto = { path = "src/scraping_coto" }

# [tool.taskipy.tasks]
# databricks-bundle = "python -m dotenv run -- databricks bundle"
# databricks-secrets = "python -m dotenv run -- databricks secrets"
# bundle-validate = "python -m dotenv run -- databricks bundle validate"
# bundle-deploy = "python -m dotenv run -- databricks bundle deploy"
# bundle-deploy-dev = "python -m dotenv run -- databricks bundle deploy --target dev"

[tool.logfire]
project_name = "dab-scraping-coto"
config_dir = "/.logfire"
